Moved the WindowResizeHandler from RenderGraph into it'w own header and source, and added the WindowResizeHandler as a member of RenderGraph to allow it to be assigned/controlled by users. Changed vsg::Context to inherit from vsg::Inherit and how the standard support for allocating on heap and ref counting.
